Keep tab active when reloading in container
Currently when one uses a shortcut to reload the current tab in a
different container, it will loose it's active status. I.e. instead of
reloading the current tab and keeping it active, it will change the
active tab to a previous tab.

This is caused by the `discarded: true` flag, which forces the reloaded
tab to be loaded in discarded state (and thus inactive state).

By removing this flag it'll now reload the tab in the new container and
keep this tab as the active tab.